Moved some function comments in scrollserver.py to docstrings.
Added the config.py file for parsing configuration options. This also adds the
'--help' commandline parameter. Some globals are now configuration parameters
to save having to edit the sourcefiles.

Changed the RequestHandler class to derive directly from
BaseHTTPRequestHandler, since this allows us to more easily set the interface
and port. It will also make life easier when it comes time to add
multi-threading or forking to the model. This required the creation of do_GET
and do_HEAD methods, but they aren't too involved.