Added
>Start
"Shareable" files are those that can be stored on one host and used on others. "Unshareable" files are those that
are not shareable. For example, the files in user home directories are shareable whereas device lock files are not.
"Static" files include binaries, libraries, documentation files and other files that do not change without system
administrator intervention. "Variable" files are files that are not static.
>End

in accordance with changes made via FHS 2.3