Added a line to create a boot disk ISO in the ZIP drive.
Modernized sub-shell calls, a few other tweaks.

Fedora Core 4 mods. Name of the directory to be saved has to be last. Also, we now specify --numeric-owner so as to avoid UID problems when using some live CD systems. And we now save to /var instead of a mounted ZIP disk.