DocBook XML DOCTYPE declaration correction
The DOCTYPE specification was missing the System Identifier
"http://www.oasis-open.org/docbook/xml/4.1.2/docbookx.dtd"
Processing is still not validating properly, because the toolchain used to
generate this, before, created the index.xml file on a first pass, and then
the actual document on the second pass, in the style of SGML tools.
I would like to figure out how to support both with a processing instruction.