DocBook XML correction
Though it is legal to define parameter entity substitutions in a
DTD's internal subset, you cannot actually use them there.  It is only
legal to use a parameter entity substitution in an external subset.

See this:
https://www.w3.org/TR/1998/REC-xml-19980210#wfc-PEinInternalSubset

Therefore, for the XML-RPC-HOWTO, the %GoodStyleSheets; definition is
left uncommented, but the suppression of the alternate definition of
legal.notice is commented out.

Document validates and processes correctly, now, using the legal.notice
definition intended.  N.B., there is no actual difference in the license
specified, just the markup used to communicate it.