rename the local index.sgml to slovak-index.sgml
When generating a book (or article) index, the filename is index.sgml.  Not an
issue with DocBook XML, only with the older tools which make several passes
over the input sources to create the index data (output as SGML in a file
called index.sgml) and then incorporate that into the final document.