allow automated publishing; switch to ./resources/
The additional programs and content were stored in a directory called
ncurses_programs and were referred to in the document--the problem is that
only ./images/ were copied to the output tree, so HTML versions of the
document would fail in building AND the files would not be viewable.

Adjusted that be creating a ./resources/ directory along the same line
as the ./images/ directory.  This can be changed, if desired, but this
allows for automated publication of the document.  (Side benefit: this is
generalizable to all other TLDP documents.)