Solve XSLT / FOP interaction bug on qandaset
The qandaset had a defaultlabel="none" attribute.  This is DocBook legal, but
the XSLT layer was producing FO output that included an empty
fo:list-item-label, with the following error message:
  "fo:list-item-label" is missing child elements. Required content model:
  marker* (%block;)+
By omitting the defaultlabel="none", the entire problem disappears.

Also, adjusted paths for reference to the ./Annimals/ which are now in
./resources/Annimals/.