removed inappropriate use of <xref endterm=""/>
This document used an endterm in the wrong place and its presence was
confusing the heck out of the FO processor:

  <xref endterm="partitiontable" linkend="partitiontable" />

The endterm="" attribute tells the DocBook processor to copy the content found
at the linkend inline.  The problem is that the content at this particular
linkend was an entire table.  This meant that the FO processor was receiving
an entire (DocBook) table (as FO, of course) and futilely trying to render it
inline.

Removing the endterm="partitiontable" allows the document to be processed by
FOP into a PDF.