initial commit; link from old tree into new one
The LDP output tree has an inconsistent structure for handling documents of
different types.  With the new processing tools, we will be able to produce
a single output directory for each source document.  This script allows
the creation of links in the historical LDP output directory to the new
automatically created (and updated) directory.